Parametrize time tests for archive and time format
The zip format used to test the time (atime, mtime, ctime) functionality
does not support high resolution time data. We need to be able to add
and test the entry_Xtime_nsec functionality for PAX tar, so a bit of
refactoring is needed to make that possible.

This refactoring makes the archive format and time format a test
parameter that can be changed. In subsequent changes, we can add in
other archive and time formats to be tested.