Support nanosecond time formats
In PAX tar, it records the nanosecond timestamp in a header. This allows
us to get the full resolution of the original timestamp back in the
archive_entry.

Similar to the libarchive code itself, we'll return the original int
time when there are no nanoseconds set. For formats where they were set,
we'll return a float timestamp to match Python handling in os.stat() and
tarfile.